refacture: moved text listing output into own module
